221102 로컬 테스트 최종

+++ proj/views/post_views.py
@@ -8,12 +8,16 @@ bp = Blueprint('post', __name__, url_prefix='/post')
 
 @bp.route('/<post>', methods=["GET","POST"])
 def home(post):
+    if session.get('id') is None:
+        return render_template('login.html')
     post_pk = post
     return render_template('post.html', post_pk = post_pk)
 
 # 포스트 불러오기
 @bp.route("/get", methods=["GET"])
 def post_get():
+    if session.get('id') is None:
+        return render_template('login.html')
     postwrite_pk_receive = request.args.get('postwrite_pk1_give')
     postwrite_list = list(db.postwrite.find({'postwrite_pk': int(postwrite_pk_receive)}, {'_id': False}))
 
@@ -22,11 +26,12 @@ def post_get():
 # 댓글 데이터 저장
 @bp.route('/comment', methods=["POST"])
 def comment():
+    if session.get('id') is None:
+        return render_template('login.html')
+
     postwrite_pk_receive = request.form['postwrite_pk_give']
     comments_receive = request.form['comments_give']
     comments_date_receive = request.form['comments_date_give']
-
-    session['id'] = 'minho080'
 
     # pk용 카운트 만들기
     box = list(db.counts.find({}))
@@ -49,6 +54,9 @@ def comment():
 # 댓글 불러오기
 @bp.route("/comment", methods=["GET"])
 def comment_get():
+    if session.get('id') is None:
+        return render_template('login.html')
+
     postwrite_pk_receive = request.args.get('postwrite_pk1_give')
     comment_list = list(db.comment.find({'postwrite_pk': int(postwrite_pk_receive)}, {'_id': False}))
 
@@ -57,6 +65,9 @@ def comment_get():
 # 댓글 갯수 표시
 @bp.route("/comment_count", methods=["GET"])
 def comment_count_get():
+    if session.get('id') is None:
+        return render_template('login.html')
+
     postwrite_pk_receive = request.args.get('postwrite_pk1_give')
     comment_list = list(db.comment.find({'postwrite_pk': int(postwrite_pk_receive)}, {'_id': False}))
 
@@ -66,11 +77,11 @@ def comment_count_get():
 # 댓글 삭제
 @bp.route('/comment/delete', methods=["POST"])
 def comment_delete():
+    if session.get('id') is None:
+        return render_template('login.html')
+
     postwrite_pk_receive = request.form['postwrite_pk_give']
     comment_pk_receive = request.form['comment_pk_give']
-
-    session['id'] = 'minho080'
-
 
     # 포스트작성자권한
     postwrite_pk1 = db.postwrite.find_one({'postwrite_pk': int(postwrite_pk_receive)}, {'_id': False})
@@ -91,9 +102,10 @@ def comment_delete():
 # 좋아요 데이터 저장
 @bp.route('/like', methods=["POST"])
 def like():
-    postwrite_pk_receive = request.form['postwrite_pk_give']
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] = 'minho080'
+    postwrite_pk_receive = request.form['postwrite_pk_give']
 
     # pk용 카운트 만들기, 좋아요 데이터 저장
     box = list(db.counts.find({}))
@@ -119,9 +131,10 @@ def like():
 # 좋아요 눌럿으면 채운하트 아니면 빈하트
 @bp.route("/like", methods=["GET"])
 def like_get():
-    postwrite_pk_receive = request.args.get('postwrite_pk1_give')
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] = 'minho080'
+    postwrite_pk_receive = request.args.get('postwrite_pk1_give')
 
     if db.like.find_one({'like_id': session['id'],'postwrite_pk': int(postwrite_pk_receive)}):
         return jsonify({'result': 1})
@@ -131,6 +144,9 @@ def like_get():
 # 좋아요 갯수 표시
 @bp.route("/like/count", methods=["GET"])
 def like_count_get():
+    if session.get('id') is None:
+        return render_template('login.html')
+
     postwrite_pk_receive = request.args.get('postwrite_pk1_give')
     like_list = list(db.like.find({'postwrite_pk': int(postwrite_pk_receive)}, {'_id': False}))
     like_count = len(like_list)
@@ -140,9 +156,10 @@ def like_count_get():
 # 좋아요 취소
 @bp.route('/like/delete', methods=["POST"])
 def like_delete():
-    postwrite_pk_receive = request.form['postwrite_pk_give']
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] = 'minho080'
+    postwrite_pk_receive = request.form['postwrite_pk_give']
 
     db.like.delete_one({'like_id': session['id'],'postwrite_pk': int(postwrite_pk_receive)})
 
@@ -154,9 +171,10 @@ def like_delete():
 # 북마크 데이터 저장
 @bp.route('/bookmark', methods=["POST"])
 def bookmark():
-    postwrite_pk_receive = request.form['postwrite_pk_give']
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] = 'minho080'
+    postwrite_pk_receive = request.form['postwrite_pk_give']
 
     # pk용 카운트 만들기
     box = list(db.counts.find({}))
@@ -176,9 +194,10 @@ def bookmark():
 # 북마크 눌렀으면 채운 북마크 아니면 빈 북마크
 @bp.route("/bookmark", methods=["GET"])
 def bookmark_get():
-    postwrite_pk_receive = request.args.get('postwrite_pk1_give')
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] ='minho080'
+    postwrite_pk_receive = request.args.get('postwrite_pk1_give')
 
     if db.bookmark.find_one({'bookmark_id': session['id'], 'postwrite_pk': int(postwrite_pk_receive)}):
         return jsonify({'result': 1})
@@ -187,9 +206,10 @@ def bookmark_get():
 # 북마크 취소
 @bp.route('/bookmark/delete', methods=["POST"])
 def bookmark_delete():
-    postwrite_pk_receive = request.form['postwrite_pk_give']
+    if session.get('id') is None:
+        return render_template('login.html')
 
-    session['id'] = 'minho080'
+    postwrite_pk_receive = request.form['postwrite_pk_give']
 
     db.bookmark.delete_one({'bookmark_id': session['id'],'postwrite_pk':int(postwrite_pk_receive)})
     return jsonify({'msg': '북마크 취소!'})
